refactor(fuse): dedupe read prefetch paths and extract poll module
- Create crates/fuse/src/poll.rs with PollResult enum (any fuse,winfsp) and
  poll_filepointer_resolution (fuse-only, pub(crate)); read_ops.rs imports them
- Remove duplicate PollResult + poll_filepointer_resolution from read_ops.rs
- Add spawn_content_prefetch_fuse helper in read_ops.rs to dedupe 3x identical
  prefetch-spawn blocks (handle_open read path + 2x handle_read sites)
- Create platform/windows/content_fetch.rs with spawn_content_prefetch (winfsp-only)
  deduping 2x prefetch closure in platform/windows/read_ops.rs
- Declare mod content_fetch in platform/windows/mod.rs
- handle_release stays in read_ops.rs (CR-04/D-04 journal-fsync-before-ack)
- cargo test -p cipherbox-fuse: 64 passed
- Windows winfsp side verified by inspection and CI (cannot compile on macOS)
